requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
google-genai>=0.1.0